Changes to sync script and new script for moving tasks across project lanes.
